Enable SSL certificate verification by default
This will require you to store a copy of the relevant chain
of certificates in a subdirectory of the config dir named
'ca-certs'.

With an OpenSSL derived version of libcurl, you will need
to run the openssl tool 'c_rehash' on 'ca-certs' before
your certificates will validate successfully.

Lastly, and not recommended, this feature may be disabled
by adding '--disable-ssl-verify' to the command.